A quick video showing changing pivot table layout
Changing the pivot table layout so that is is more legible for people to read and takes up less space. Here the goal is to display the total sales of chocolates based on different geographical locations and the first choice of all three criteria/ labels appearing as rows creates too many rows and is very confusing to view. Now with the "Geography" criteria/parameter being represented via column instead of row the pivot table will be much more readable.