Pivot tables making top 10 product easier
I have done this scenario with other data before via using VBA formulas/code used in Excel to find the top 10 of a product or top 10 in sales for a company etc. And it is much easier with pivot tables. Glad I could see first hand how amazing these pivot tables are compared to the more hands on coding approach that took me much longer to learn and practice the appropriate formulas because often times I had to use nested formulas that often confused me or had a syntax error that caused the formulas not to return the expected values. Pivot tables are the way to go, no wonder so many job postings require people to use them because they are a time-saver and stop you from performing added layers of complexity. Not to mention the simple drop-down interface makes it so easy to share the data and what you have found with anyone including those that may not be able to come up with the formulas on their own. Truly a wonderful process to learn these pivot tables because they are fairly intuitive to learn and save you a lot of head scratching!