Pivot tables creating advanced filtering the easy way
Without pivot tables I would have had to do complicated nesting of formulas and would have taken at least 10 minutes to figure out the correct syntax of how to find the data that is need to filter out a specific salesman or saleswoman and their specific top ten chocolate products they sold and the associated number of customers. These pivot tables are simply create a layer of abstraction for the end user that is very efficient in getting large amounts of data analyzed and filtered in a visually appealing way. I say abstraction because you don't need to know how these pivot tables function behind the scenes to accomplish the task at hand, you just need to know what you are looking for and the pivot tables do the work for you.